requirements.txt
asgiref==3.10.0
Django==5.2.7
django-axes==8.0.0
django-debug-toolbar==6.0.0
django-phonenumber-field==8.3.0
django-stubs==5.2.7
django-stubs-ext==5.2.7
djangorestframework==3.16.1
phonenumbers==9.0.17
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
tzdata==2025.2
